Add BAZEL_BUILD_DIRECTORY option
It is often useful to limit the Docker build context to a specific
directory to reduce the number of files sent to the daemon and speed
up builds.

Add a BAZEL_BUILD_DIRECTORY option to specify the directory to build
the Docker image in, which does not have to be identical to the
workspace root.